validated the whole cycle
