+++ dspot-prettifier/src/main/java/eu/stamp_project/prettifier/context2name/C2N/python/c2n_apply.py
@@ -36,7 +36,7 @@ class Config:
 
 
 def easy_path(file_path):
-    return f"model/{file_path}"
+    return f"../model/{file_path}"
 
 
 def easy_open(file_path, mode):
